fixed cinder issue and disabled sensu
This commit contains fix for cinder's auth_uri and identity_uri
to be set properly by puppet. The changes also include disabling
the sensu from being installed on the nodes. It also contains
changes to modify /etc/hosts file. The change of /etc/hosts is
still in testing and might get changed to something else in
future.